server/requirements.txt
sanic == 21.9.1
pillow == 6.2.2
pandas == 1.3.4
scikit-learn ==  1.0.1